+++ tests/client/test_search.py
@@ -1,6 +1,8 @@
-from pubtools.pulplib import Criteria
+import pytest
 
-from pubtools.pulplib._impl.client.search import filters_for_criteria
+from pubtools.pulplib import Criteria, Matcher
+
+from pubtools.pulplib._impl.client.search import filters_for_criteria, field_match
 
 
 def test_null_criteria():
@@ -46,3 +48,24 @@ def test_field_or_criteria():
     assert filters_for_criteria(Criteria.or_(c1, c2)) == {
         "$or": [{"f1": {"$eq": "v1"}}, {"f2": {"$eq": "v2"}}]
     }
+
+
+def test_field_regex_criteria():
+    """with_field with regex is translated to a mongo fragment as expected."""
+
+    assert filters_for_criteria(
+        Criteria.with_field("some.field", Matcher.regex("abc"))
+    ) == {"some.field": {"$regex": "abc"}}
+
+
+def test_non_matcher():
+    """field_match raises if invoked with something other than a matcher.
+
+    This is not possible to reach by public API alone; it'd be an internal bug
+    if this exception is ever raised.
+    """
+
+    with pytest.raises(TypeError) as exc_info:
+        field_match("oops not a matcher")
+
+    assert "Not a matcher" in str(exc_info)